Reload row summary when viewing list of users
When others have changed their responses, the page gets
stale, showing an incorrect summary count of each response
type. As a result, clicking on the "Who" button for a time may
show a list of names that does not match the count displayed.

As a stopgap measure, this commit reloads just the time
that was clicked on to get the latest count along with the
current player list.

A future feature should use a push-update model to keep
the web page updated for all events, live, as people respond.